requirements.txt
uvicorn
fastapi

python-dotenv
sqlalchemy

pydantic